# src/app/api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "PSL Studio API",
    "description": "REST API for PSL Studio — a fashion product lifecycle management platform covering products, samples, events, campaigns, loans, media placements, and guests.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development server"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token obtained from POST /api/auth/login or POST /api/auth/register. Pass in the Authorization header as: Bearer <token>"
      }
    },
    "schemas": {
      "Pagination": {
        "type": "object",
        "properties": {
          "page": { "type": "integer", "example": 1 },
          "limit": { "type": "integer", "example": 20 },
          "total": { "type": "integer", "example": 100 },
          "totalPages": { "type": "integer", "example": 5 }
        }
      },
      "Error": {
        "type": "object",
        "properties": {
          "error": { "type": "string" }
        },
        "required": ["error"]
      },
      "Success": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean", "example": true }
        }
      },
...
      "Event": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" },
          "deletedAt": { "type": "string", "format": "date-time", "nullable": true },
          "profileId": { "type": "string" },
          "name": { "type": "string" },
          "description": { "type": "string", "nullable": true },
          "type": {
            "type": "string",
            "enum": ["SHOW", "PRESENTATION", "LAUNCH", "PRESS", "TRADE-SHOW", "OTHER"]
          },
          "status": {
            "type": "string",
            "enum": ["DRAFT", "CONFIRMED", "COMPLETED", "CANCELLED"]
          },
          "startAt": { "type": "string", "format": "date-time" },
          "endAt": { "type": "string", "format": "date-time", "nullable": true },
          "location": { "type": "string", "nullable": true },
          "venue": { "type": "string", "nullable": true }
        }
      },
      "EventCreate": {
        "type": "object",
        "required": ["name", "type", "startAt"],
        "properties": {
          "name": { "type": "string", "minLength": 1, "maxLength": 200 },
          "description": { "type": "string", "maxLength": 2000, "nullable": true },
          "type": {
            "type": "string",
            "enum": ["SHOW", "PRESENTATION", "LAUNCH", "PRESS", "TRADE-SHOW", "OTHER"]
          },
          "status": {
            "type": "string",
            "enum": ["DRAFT", "CONFIRMED", "COMPLETED", "CANCELLED"],
            "default": "DRAFT"
          },
          "startAt": { "type": "string", "format": "date-time" },
          "endAt": { "type": "string", "format": "date-time", "nullable": true },
          "location": { "type": "string", "maxLength": 500, "nullable": true },
          "venue": { "type": "string", "maxLength": 300, "nullable": true }
        }
      },
...
      "EventProductBody": {
        "type": "object",
        "required": ["productId"],
        "properties": {
          "productId": { "type": "string" },
          "notes": { "type": "string", "nullable": true },
          "look": { "type": "integer", "nullable": true, "description": "Look number for fashion shows" }
        }
      },
      "EventProductDelete": {
        "type": "object",
        "required": ["productId"],
        "properties": {
          "productId": { "type": "string" }
        }
      },
...
    },
    "parameters": {
      "idParam": {
        "name": "id",
        "in": "path",
        "required": true,
        "schema": { "type": "string" },
        "description": "Resource identifier (cuid)"
      },
      "pageParam": {
        "name": "page",
        "in": "query",
        "schema": { "type": "integer", "minimum": 1, "default": 1 },
        "description": "Page number (1-based)"
      },
      "limitParam": {
        "name": "limit",
        "in": "query",
        "schema": { "type": "integer", "minimum": 1, "maximum": 100, "default": 20 },
        "description": "Number of items per page"
      }
    },
    "responses": {
      "Unauthorized": {
        "description": "Missing or invalid authentication token",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Error" },
            "example": { "error": "Non authentifié" }
          }
        }
      },
...
      "NotFound": {
        "description": "Resource not found",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Error" }
          }
        }
      },
      "BadRequest": {
        "description": "Invalid request body or parameters",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Error" }
          }
        }
      },
...
    }
  },
...
  "paths": {
...
    "/api/events": {
      "get": {
        "tags": ["Events"],
        "summary": "List events (paginated)",
        "operationId": "listEvents",
        "parameters": [
          { "$ref": "#/components/parameters/pageParam" },
          { "$ref": "#/components/parameters/limitParam" },
          {
            "name": "status",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": ["DRAFT", "CONFIRMED", "COMPLETED", "CANCELLED"]
            },
            "description": "Filter by event status"
          },
          {
            "name": "type",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": ["SHOW", "PRESENTATION", "LAUNCH", "PRESS", "TRADE-SHOW", "OTHER"]
            },
            "description": "Filter by event type"
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated list of events",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "allOf": [
                          { "$ref": "#/components/schemas/Event" },
                          {
                            "type": "object",
                            "properties": {
                              "_count": {
                                "type": "object",
                                "properties": {
                                  "campaigns": { "type": "integer" },
                                  "products": { "type": "integer" }
                                }
                              }
                            }
                          }
                        ]
                      }
                    },
                    "pagination": { "$ref": "#/components/schemas/Pagination" }
                  }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      },
      "post": {
        "tags": ["Events"],
        "summary": "Create a new event",
        "operationId": "createEvent",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/EventCreate" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Event created",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Event" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
...
    "/api/events/{id}/products": {
      "post": {
        "tags": ["Events"],
        "summary": "Add or update a product in an event (upsert)",
        "operationId": "addEventProduct",
        "parameters": [
          { "$ref": "#/components/parameters/idParam" }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/EventProductBody" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Product added to event",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": { "type": "string" },
                    "eventId": { "type": "string" },
                    "productId": { "type": "string" },
                    "addedAt": { "type": "string", "format": "date-time" },
                    "notes": { "type": "string", "nullable": true },
                    "look": { "type": "integer", "nullable": true }
                  }
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      },
      "delete": {
        "tags": ["Events"],
        "summary": "Remove a product from an event",
        "operationId": "removeEventProduct",
        "parameters": [
          { "$ref": "#/components/parameters/idParam" }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/EventProductDelete" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Product removed from event",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Success" }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
...
    "/api/placements/{id}": {
      "delete": {
        "tags": ["Placements"],
        "summary": "Delete a media placement",
        "operationId": "deletePlacement",
        "parameters": [
          { "$ref": "#/components/parameters/idParam" }
        ],
        "responses": {
          "200": {
            "description": "Placement deleted",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Success" }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
...
  }
}